job_portal/requirements/base.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pytz==2024.2
sqlparse==0.5.2
tzdata==2024.2
whitenoise
